backend/requirements.txt
flask
flask-cors
paho-mqtt
pandas
scikit-learn
psycopg2-binary
gunicorn